Subprocess Editing
There was some issues when switching between Windows and Ubuntu that the DARP.java and ConnectComponent.java in the darp_pack had to be recompiled.
Those compilations are now included in the subprocess file. It might be useful to have a separate subprocess file that runs once, but not sure how to implement that.