Finalised Prim
Final Prim MST created. The printing bug on the original DARP area division graph was fixed (Y-vals start from 0 and robots are on grid).

Next is the paths to circumvent the trees. Starting point is creating a grid double the size of the original to store the values for the environment.