Robot Initial Positions Shift
Finalised shifting the robot initial positions by one half cell - not as trivial as it sounds. Took a whole day.
Not sure how to establish these robot positions later with the refuelling. At present they start at cell centres and are then moved to the boundaries. Moving to the boundaries depends on the direction the robot is moving - hence not pre-emptively placing robots on boundaries and hence the code is not trivial.