DARP alteration - Manhattan Distance
Added Manhattan Distance Capability to DARP algorithm (Should include a variable that picks which one to use as input). Not sure if the normalization remains the same. It works without changing the normalization.

It may run faster when using Manhattan distance. Should consider adding the possibility of Geodesic Manhattan like in the Voronoi Paper. Will probably need to use A* which could be too processor intensive. Depends on how much it improves DARP's time to completion.