Target Finding
It can locate the target and stop the path appropriately for the robot that finds it.

Still need to add ability for the other paths to stop at the same time as well.

Still need to implement shortest path back to starting position (for now -this can be adapted later so that the drones perform some other function after target finding).