Tangent-Line Path Planner Finished
Finished porting the path planner into python. Going to use it with the refueling protocol, to ascertain the minimum path to starting point for each UAV (for fixed wings). For VTOL and multi-rotors the paths will look different.

Can also be used for landing. Just need to do some math to get the inputs to the algorithm. Outputs should be path length and corresponding time to take-off.

Should get some kind of average time to take-off and reach appropriate speed (in a straight line). The end of this line is then the starting point to the point-to-point planner.

This planner does not avoid obstacles.