Final Waypoint Generation Patch Up
Fixed the final waypoint generation in the prim algorithm. It no longer generates double waypoints and distance, waypoint and time final arrays are all the same size.

This allows for easily generating time snapshots of the final waypoints on the graph in the print graph function. This function is also now modular and is no longer part of the waypoint generation.

Target being found in later refuel runs is also now accounted for.

Next is adding landing and take off times using Ian Peddle's Planner. This will be used to implement scheduling.